+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Actor/ActorData.cpp
@@ -103,6 +103,9 @@ void FVehicleData::RecordActorData(FCarlaActor* CarlaActor, UCarlaEpisode* Carla
     PhysicsControl = Vehicle->GetVehiclePhysicsControl();
   }
   Control = Vehicle->GetVehicleControl();
+  AckermannControl = Vehicle->GetVehicleAckermannControl();
+  bAckermannControlActive = Vehicle->IsAckermannControlActive();
+  AckermannControllerSettings = Vehicle->GetAckermannControllerSettings();
   LightState = Vehicle->GetVehicleLightState();
   auto Controller = Cast<AWheeledVehicleAIController>(Vehicle->GetController());
   if (Controller)
@@ -121,7 +124,15 @@ void FVehicleData::RestoreActorData(FCarlaActor* CarlaActor, UCarlaEpisode* Carl
   {
     Vehicle->ApplyVehiclePhysicsControl(PhysicsControl);
   }
-  Vehicle->ApplyVehicleControl(Control, EVehicleInputPriority::Client);
+  Vehicle->ApplyAckermannControllerSettings(AckermannControllerSettings);
+  if (!bAckermannControlActive)
+  {
+    Vehicle->ApplyVehicleControl(Control, EVehicleInputPriority::Client);
+  }
+  else 
+  {
+    Vehicle->ApplyVehicleAckermannControl(AckermannControl, EVehicleInputPriority::Client);
+  }
   Vehicle->SetVehicleLightState(LightState);
   auto Controller = Cast<AWheeledVehicleAIController>(Vehicle->GetController());
   if (Controller)