+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ## Latest
 
   * Use MediaCapture to capture the scene for the camera RGB sensor. It speeds up the FPS
+  * Added native ackermann controller:
+    - `apply_ackermann_control`:  to apply an ackermann control command to a vehicle
+    - `get_ackermann_controller_settings`: to get the last ackermann controller settings applied
+    - `apply_ackermann_controller_settings`: to apply new ackermann controller settings
+  * Fixed bug causing the Traffic Manager to not be deterministic when using hybrid mode
+  * Added `NormalsSensor`, a new sensor with normals information
+  * Added support for N wheeled vehicles
 
 ## CARLA 0.9.13